moving simple_sar domain to domains folder
